refactor: move code handling request to controller
I also rearrange the order of each function in chatRouter to original one
