+++ backend/routes/chatRouter.js
@@ -1,70 +1,8 @@
 const express = require("express");
 const path = require("path");
-const { create_message, subscribe_to_chat } = require("../controllers/chatController");
+const { create_message, subscribe_to_chat, join_chat, create_chat, create_message } = require("../controllers/chatController");
 const chatRouter = express.Router();
 const Chat = require("../models/Chat.js");
-
-// Join a chat
-chatRouter.get("/join/:chat_id", async (req, res) => {
-	console.log("GET request recieved");
-    const chatId = req.params.chat_id;
-
-    try {
-		console.log("Attempting to find chat");
-        const chat = await Chat.findOne({ pin: chatId });
-		console.log("Chat found");
-
-        if (chat) {
-			res.json({ exists: true });
-        } else {
-            res.status(404).json({ exists: false, message: "Chatroom not found." });
-        }
-    } catch (error) {
-        console.error("Error checking chatroom existence:", error);
-        res.status(500).json({ message: "An error occurred while checking the chatroom." });
-    }
-});
-
-//Utility function to generate PIN
-function generateRandomPin() {
-    return Math.floor(10000000 + Math.random() * 90000000);
-}
-
-// Create a new chat
-chatRouter.post("/create", async (req, res) => {
-	console.log("POST request recieved");
-    const pin = generateRandomPin();  // Generate a random PIN
-    const chatName = "Anonymous Chat";  // Set a default chat name
-
-    try {
-		console.log("Attempting to create new Chat")
-        // Create a new chat instance
-        const newChat = new Chat({
-            name: chatName,
-            pin: pin,
-            users: [],  // No users initially
-            message: [], // No messages initially
-        });
-		console.log("Chat created successfully")
-
-        // Save the chat to the database
-        try {
-			// Attempt to save the new chat instance to the database
-			await newChat.save();
-			console.log("Chat saved successfully");
-
-			// Send a response after saving the chat
-			res.json({ chatId: pin });
-		} catch (error) {
-			// Log the error details to understand what went wrong
-			console.error("Error saving chat:", error);
-			return res.status(500).json({ message: "An error occurred while creating the chat" });
-		}
-    } catch (error) {
-        console.error("Error creating chat:", error);
-        return res.status(500).json({ message: "An error occurred while creating the chat" });
-    }
-});
 
 // Get a chat
 chatRouter.get("/:chat_id", (req, res) => {
@@ -78,6 +16,12 @@ chatRouter.get("/:chat_id/events", subscribe_to_chat);
 chatRouter.get("/:chat_id/:poll_id", (req, res) => {
     res.send(`Poll ID: ${req.params.poll_id}`);
 });
+
+// Join a chat
+chatRouter.get("/join/:chat_id", join_chat);
+
+// Create a new chat
+chatRouter.post("/create", create_chat);
 
 // Post a message to a chat
 chatRouter.post("/:chat_id/:user_id/:message_content", create_message);
